move checklists out of flora folder
- redirects have been put in place for old routes
